// src/utils/interfaces.ts
interface ApiMatch {
  away_goal: number;
  away_team_name: string;
  away_team_short_name: string;
  category_status: string;
  championship_title: string;
  date_time: string;
  home_goal: number;
  home_team_name: string;
  home_team_short_name: string;
  match_day_category_status: string;
  match_day_title: string;
  match_hm: string;
  round_title: string;
}

export { ApiMatch };